custom interval & ajax to check logged in state
Removed all use of Heartbeat API since it is blocked by WP Engine (and
maybe other hosts) on front-end. Set interval to 5 minutes. Also
removed all code that logged user out as we are using “Prevent
Concurrent Logins” plugin for that. This plugin is now just the
front-end periodic check to see if a logged in session is still active.
Also renamed plugin based on new functionality.